WIP finish arroy and start qdrant
